refactor(cockpit): extract display helpers
